% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{db_con}
\alias{db_con}
\title{Connect to the intendo database using stored credentials}
\usage{
db_con()
}
\description{
With the appropriate environment variables in place (ask your manager for
access to these values), this function simply creates a connection object.
}
\details{
If you don't have the environment variables set (but have the values) the
easiest way to do this is to use the \code{usethis::edit_r_environ()} function.
That puts you into your \code{.REnviron} file. You need to have the \code{DBNAME},
\code{USERNAME}, \code{PASSWORD}, \code{HOST}, and \code{PORT} environment variables set with the
appropriate values. If you don't wish to store the DB password on your
system, you can use the \code{\link[=db_con_p]{db_con_p()}} function (which asks for the password
with every use but requires every other environment variable to be available).
}
