% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_kpis.R
\name{main_kpi}
\alias{main_kpi}
\alias{get_dau}
\alias{get_mau}
\alias{get_dac}
\alias{get_arpu}
\title{Functions that can help us get the DAU, MAU, DAC, and ARPU over a set of dates}
\usage{
get_dau(con, dates)

get_mau(con, year, month)

get_dac(con, dates)

get_arpu(con, dates)
}
\arguments{
\item{con}{a database connection}

\item{dates}{a set of dates}
}
\value{

}
\description{
Functions that can help us get the DAU, MAU, DAC, and ARPU over a set of dates
}
\examples{
library(intendo)
intendo <- db_con()
# Test the `get_dau()` function
get_dau(con = intendo, dates = c("2015-01-01", "2015-01-02", "2015-01-03"))
get_mau(con = intendo, dates = c("2015-01-01", "2015-01-02", "2015-01-03"))

}
